+++ Final Project/src/com/chickenachiever/model/USpike.java
@@ -12,21 +12,28 @@ public class USpike extends MapElementRevamp {
 	private ArrayList<BufferedImage[]> sprites;
 	private final int[] numFrames = { 1 };
 	private static final int IDLE = 0;
+	private PlayerRevamp p;
 
 	public USpike(TileMapRevamp tm, int x, int y) {
-		super(tm,x,y);
+		super(tm, x, y);
 		cwidth = 10;
 		cheight = 10;
 	}
 
+	public void setPlayer(PlayerRevamp p) {
+		this.p = p;
+	}
 
-	public void update(PlayerRevamp p) {
-		if(intersects(p)){
-			p.kill();
+	public void update() {
+		//System.out.println("hi");
+		if (p != null) {
+			if (intersects(p)) {
+				p.kill();
+			}
 		}
 	}
 
 	public void draw(Graphics2D g2d) {
 	}
-	
+
 }
